upgrade dip20 standard to use thread_local main branch
