+++ ic/weth/src/weth/weth.did
@@ -1,16 +1,3 @@
-type Result = variant { Ok : nat; Err : TxError };
-type TxError = variant {
-  InsufficientAllowance;
-  InsufficientBalance;
-  Canister : text;
-  ErrorOperationStyle;
-  Unauthorized;
-  LedgerTrap;
-  ErrorTo;
-  Other;
-  BlockUsed;
-  AmountTooSmall;
-};
 type Metadata = record {
   fee : nat;
   decimals : nat8;
@@ -39,6 +26,7 @@ type TxError = variant {
   Other;
   BlockUsed;
   AmountTooSmall;
+  Canister : text;
 };
 service : (
   text,
@@ -75,7 +63,7 @@ service : (
   totalSupply : () -> (nat) query;
   transfer : (principal, nat) -> (Result);
   transferFrom : (principal, principal, nat) -> (Result);
-  
+
   // proxy
   burn : (principal, nat) -> (Result);
   mint : (nat, vec nat) -> (Result);
